Move mod output code to output module
