+++ Vic2ToHoI4/Source/OutHoi4/OutMod.h
@@ -1,0 +1,15 @@
+#ifndef OUT_MOD_H
+#define OUT_MOD_H
+
+
+
+#include "../HOI4World/HoI4World.h"
+
+
+
+void clearOutputFolder();
+void output(HoI4::World& destWorld);
+
+
+
+#endif // OUT_MOD_H
