+++ Vic2ToHoI4Tests/Vic2ToHoI4Tests-Google.vcxproj.filters
@@ -851,6 +851,9 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\OutHoi4\OutHoi4World.cpp">
       <Filter>Vic2ToHoI4 files\OutHoi4</Filter>
     </ClCompile>
+    <ClCompile Include="..\Vic2ToHoI4\Source\OutHoi4\OutMod.cpp">
+      <Filter>Vic2ToHoI4 files\OutHoi4</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Vic2ToHoI4 files">
@@ -1076,6 +1079,9 @@
     </ClInclude>
     <ClInclude Include="Mocks\CoastalProvincesMock.h">
       <Filter>Mocks</Filter>
+    </ClInclude>
+    <ClInclude Include="..\Vic2ToHoI4\Source\OutHoi4\OutMod.h">
+      <Filter>Vic2ToHoI4 files\OutHoi4</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
